requirements.txt
# Requirements para desenvolvimento local
google-cloud-storage==2.10.0
google-cloud-vision==3.5.0
google-cloud-firestore==2.11.0
google-cloud-pubsub==2.18.0
flask==2.3.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
